+++ src/Data/RelationshipData.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Threls\FilamentPageBuilder\Data;
+
+use Spatie\LaravelData\Attributes\MapName;
+use Spatie\LaravelData\Data;
+use Spatie\LaravelData\Mappers\SnakeCaseMapper;
+
+#[MapName(SnakeCaseMapper::class)]
+class RelationshipData extends Data
+{
+    public function __construct(
+        public string $relationship,
+    ) {}
+}