+++ src/Enums/PageRelationshipTypeEnum.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Threls\FilamentPageBuilder\Enums;
+
+enum PageRelationshipTypeEnum: string
+{
+    case TESTIMONIAL = 'testimonial';
+    case FAQ = 'faq';
+    case EVENT = 'event';
+}