+++ src/PageBuilderPlugin.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Threls\FilamentPageBuilder;
+
+use Filament\Contracts\Plugin;
+use Filament\Panel;
+
+class PageBuilderPlugin implements Plugin
+{
+    public function getId(): string
+    {
+        return 'filament-page-builder';
+    }
+
+    public function register(Panel $panel): void
+    {
+        $panel
+            ->resources([
+                Resources\PageResource::class,
+            ]);
+    }
+
+    public function boot(Panel $panel): void
+    {
+        //
+    }
+
+    public static function make(): static
+    {
+        return app(static::class);
+    }
+
+    public static function get(): static
+    {
+        return filament(app(static::class)->getId());
+    }
+}