+++ src/PageBuilderServiceProvider.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Threls\FilamentPageBuilder;
+
+use Spatie\LaravelPackageTools\Package;
+use Spatie\LaravelPackageTools\PackageServiceProvider;
+
+class PageBuilderServiceProvider extends PackageServiceProvider
+{
+    public static string $name = 'filament-page-builder';
+
+    public static string $viewNamespace = 'filament-page-builder';
+
+    public function configurePackage(Package $package): void
+    {
+        $package->name(static::$name)
+            ->hasConfigFile()
+            ->hasViews()
+            ->hasMigration('2025_01_28_120000_create_pages_table');
+    }
+
+    public function packageRegistered(): void
+    {
+        parent::packageRegistered();
+    }
+
+    public function packageBooted(): void
+    {
+        parent::packageBooted();
+    }
+}