+++ src/Resources/PageResource/Pages/CreatePage.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Threls\FilamentPageBuilder\Resources\PageResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use Threls\FilamentPageBuilder\Resources\PageResource;
+
+class CreatePage extends CreateRecord
+{
+    protected static string $resource = PageResource::class;
+}