+++ src/Resources/PageResource/Pages/EditPage.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Threls\FilamentPageBuilder\Resources\PageResource\Pages;
+
+use Filament\Actions;
+use Filament\Resources\Pages\EditRecord;
+use Threls\FilamentPageBuilder\Resources\PageResource;
+
+class EditPage extends EditRecord
+{
+    protected static string $resource = PageResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\DeleteAction::make(),
+        ];
+    }
+}